handle request error in webhook
