+++ pkg/webhook/useridentity/mutatingWebhook_test.go
@@ -22,99 +22,85 @@ const (
 	appsubTest  = `{"apiVersion":"apps.open-cluster-management.io/v1","kind": "Subscription","metadata": {"name": "git-sub","namespace": "parentsub","annotations": {"apps.open-cluster-management.io/cluster-admin": "true","apps.open-cluster-management.io/github-path": "test/e2e/github/nestedSubscription"}},"spec": {"channel": "ch-git/git","placement": {"local": "true"}}}`
 )
 
-func newAdmissionReview() *v1.AdmissionReview {
-	return &v1.AdmissionReview{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "AdmissionReview",
-			APIVersion: "admission.k8s.io/v1",
+func newAdmissionRequest() *v1.AdmissionRequest {
+	return &v1.AdmissionRequest{
+		UID: "3f6d1d1f-61a0-49ea-b458-05454ba42ab6",
+		Kind: metav1.GroupVersionKind{
+			Group:   "apps.open-cluster-management.io",
+			Kind:    "Channel",
+			Version: "v1",
 		},
-		Request: &v1.AdmissionRequest{
-			UID: "3f6d1d1f-61a0-49ea-b458-05454ba42ab6",
-			Kind: metav1.GroupVersionKind{
-				Group:   "apps.open-cluster-management.io",
-				Kind:    "Channel",
-				Version: "v1",
-			},
-			Resource: metav1.GroupVersionResource{
-				Group:    "apps.open-cluster-management.io",
-				Version:  "v1",
-				Resource: "channels",
-			},
-			SubResource: "",
-			RequestKind: &metav1.GroupVersionKind{
-				Group:   "apps.open-cluster-management.io",
-				Kind:    "Channel",
-				Version: "v1",
-			},
-			RequestResource: &metav1.GroupVersionResource{
-				Group:    "apps.open-cluster-management.io",
-				Version:  "v1",
-				Resource: "channels",
-			},
-			RequestSubResource: "",
-			Name:               "test",
-			Namespace:          "default",
-			Operation:          "CREATE",
-			UserInfo: authenticationv1.UserInfo{
-				Username: "system:admin",
-			},
-			Object: runtime.RawExtension{
-				Raw: []byte(channelTest),
-			},
+		Resource: metav1.GroupVersionResource{
+			Group:    "apps.open-cluster-management.io",
+			Version:  "v1",
+			Resource: "channels",
 		},
-		Response: nil,
+		SubResource: "",
+		RequestKind: &metav1.GroupVersionKind{
+			Group:   "apps.open-cluster-management.io",
+			Kind:    "Channel",
+			Version: "v1",
+		},
+		RequestResource: &metav1.GroupVersionResource{
+			Group:    "apps.open-cluster-management.io",
+			Version:  "v1",
+			Resource: "channels",
+		},
+		RequestSubResource: "",
+		Name:               "test",
+		Namespace:          "default",
+		Operation:          "CREATE",
+		UserInfo: authenticationv1.UserInfo{
+			Username: "system:admin",
+		},
+		Object: runtime.RawExtension{
+			Raw: []byte(channelTest),
+		},
 	}
 }
 
-func appSubAdmissionReview() *v1.AdmissionReview {
-	return &v1.AdmissionReview{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "AdmissionReview",
-			APIVersion: "admission.k8s.io/v1",
+func appSubAdmissionRequest() *v1.AdmissionRequest {
+	return &v1.AdmissionRequest{
+		UID: "4d6d1d1f-61a0-49ea-b458-05454ba42ab6",
+		Kind: metav1.GroupVersionKind{
+			Group:   "apps.open-cluster-management.io",
+			Kind:    "Subscription",
+			Version: "v1",
 		},
-		Request: &v1.AdmissionRequest{
-			UID: "4d6d1d1f-61a0-49ea-b458-05454ba42ab6",
-			Kind: metav1.GroupVersionKind{
-				Group:   "apps.open-cluster-management.io",
-				Kind:    "Subscription",
-				Version: "v1",
-			},
-			Resource: metav1.GroupVersionResource{
-				Group:    "apps.open-cluster-management.io",
-				Version:  "v1",
-				Resource: "subscriptions",
-			},
-			SubResource: "",
-			RequestKind: &metav1.GroupVersionKind{
-				Group:   "apps.open-cluster-management.io",
-				Kind:    "Subscription",
-				Version: "v1",
-			},
-			RequestResource: &metav1.GroupVersionResource{
-				Group:    "apps.open-cluster-management.io",
-				Version:  "v1",
-				Resource: "subscriptions",
-			},
-			RequestSubResource: "",
-			Name:               "subtest",
-			Namespace:          "default",
-			Operation:          "CREATE",
-			UserInfo: authenticationv1.UserInfo{
-				Username: "system:serviceaccount:open-cluster-management-agent-addon:klusterlet-addon-appmgr",
-			},
-			Object: runtime.RawExtension{
-				Raw: []byte(appsubTest),
-			},
+		Resource: metav1.GroupVersionResource{
+			Group:    "apps.open-cluster-management.io",
+			Version:  "v1",
+			Resource: "subscriptions",
 		},
-		Response: nil,
+		SubResource: "",
+		RequestKind: &metav1.GroupVersionKind{
+			Group:   "apps.open-cluster-management.io",
+			Kind:    "Subscription",
+			Version: "v1",
+		},
+		RequestResource: &metav1.GroupVersionResource{
+			Group:    "apps.open-cluster-management.io",
+			Version:  "v1",
+			Resource: "subscriptions",
+		},
+		RequestSubResource: "",
+		Name:               "subtest",
+		Namespace:          "default",
+		Operation:          "CREATE",
+		UserInfo: authenticationv1.UserInfo{
+			Username: "system:serviceaccount:open-cluster-management-agent-addon:klusterlet-addon-appmgr",
+		},
+		Object: runtime.RawExtension{
+			Raw: []byte(appsubTest),
+		},
 	}
 }
 
 func TestMutateResource(t *testing.T) {
 	adHandler := newAdmissionHandler()
-	rsp := adHandler.mutateResource(*newAdmissionReview())
+	rsp := adHandler.mutateResource(newAdmissionRequest())
 	assert.True(t, true, len(rsp.Patch) != 0)
 
-	rsp = adHandler.mutateResource(*appSubAdmissionReview())
+	rsp = adHandler.mutateResource(appSubAdmissionRequest())
 	assert.True(t, rsp.Patch == nil)
 }
